More refactoring and added a namespace selection
