+++ shared/backend/aws/interfaces/mock_compute_backend.go
@@ -111,19 +111,19 @@ func (mr *MockComputeBackendMockRecorder) GetResources(arg0, arg1 interface{}) *
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetResources", reflect.TypeOf((*MockComputeBackend)(nil).GetResources), arg0, arg1)
 }
 
-// ListClusterIds mocks base method.
-func (m *MockComputeBackend) ListClusterIds(arg0 context.Context) ([]string, error) {
+// ListHappyNamespaces mocks base method.
+func (m *MockComputeBackend) ListHappyNamespaces(arg0 context.Context) ([]string, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ListClusterIds", arg0)
+	ret := m.ctrl.Call(m, "ListHappyNamespaces", arg0)
 	ret0, _ := ret[0].([]string)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
-// ListClusterIds indicates an expected call of ListClusterIds.
-func (mr *MockComputeBackendMockRecorder) ListClusterIds(arg0 interface{}) *gomock.Call {
+// ListHappyNamespaces indicates an expected call of ListHappyNamespaces.
+func (mr *MockComputeBackendMockRecorder) ListHappyNamespaces(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListClusterIds", reflect.TypeOf((*MockComputeBackend)(nil).ListClusterIds), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListHappyNamespaces", reflect.TypeOf((*MockComputeBackend)(nil).ListHappyNamespaces), arg0)
 }
 
 // PrintLogs mocks base method.
